Reassert /bin/sh validity
Reverts the shebang change in 055978e and adds a shellcheck directive
to suppress warnings about POSIX validity and prevent further drift.

The POSIX standard is not, despite its name, a reliable guide to what
features are actually standard, even for POSIX shells. rustup-init.sh
uses `local`, a non-POSIX extension supported by all POSIX shells known
in common use. Thus, it is fine to run it with #!/bin/sh instead.

The Debian Almquist shell is not the only POSIX shell but it is very
common, so is a useful proxy for whether the script actually runs for
non-bash shells.